alignment support
Add support for alignments in the Treebank Template. Alignments use the
`alignment-react` library: https://github.com/perseids-tools/alignment-react
and only support the Perseids alignment format.

Alignments are configured at the `section`-level of `config.json`. They
differ from a treebank only in that they:
1. Have `"type": "alignment"`
2. (Optionally) have an `"l1": <lnum>` and `"l2": <lnum>`. These are the
   language identifiers in the XML and default to `"L1"` and `"L2"`.

This PR does not include changes to the documentation because this
feature is still somewhat experimental.